Move tests to package c.y.container.plugin
