+++ bundle-plugin-test/integration-test/src/test/java/com/yahoo/container/plugin/BundleTest.java
@@ -1,5 +1,5 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo;
+package com.yahoo.container.plugin;
 
 import com.yahoo.osgi.maven.ProjectBundleClassPaths;
 import com.yahoo.vespa.config.VespaVersion;
